# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PoE2 Hideout Butler API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/callback": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "OAuth2 callback",
        "operationId": "callback_api_auth_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout and revoke session",
        "operationId": "logout_api_auth_logout_post",
        "parameters": [
          {
            "name": "poe2b_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Poe2B Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Logout Api Auth Logout Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/characters": {
      "get": {
        "tags": [
          "characters"
        ],
        "summary": "List characters for a league",
        "operationId": "list_characters_api_characters_get",
        "parameters": [
          {
            "name": "league",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "League"
            }
          },
          {
            "name": "poe2b_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Poe2B Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharactersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/characters/{name}": {
      "get": {
        "tags": [
          "characters"
        ],
        "summary": "Character detail with equipped items",
        "operationId": "get_character_api_characters__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          },
          {
            "name": "poe2b_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Poe2B Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CharacterDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/prefs": {
      "get": {
        "tags": [
          "prefs"
        ],
        "summary": "Current user preferences",
        "operationId": "get_prefs_api_prefs_get",
        "parameters": [
          {
            "name": "poe2b_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Poe2B Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrefsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "prefs"
        ],
        "summary": "Update user preferences",
        "operationId": "patch_prefs_api_prefs_patch",
        "parameters": [
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "name": "poe2b_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Poe2B Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrefsPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrefsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CharacterDetail": {
        "properties": {
          "summary": {
            "$ref": "#/components/schemas/CharacterSummary"
          },
          "equipped": {
            "items": {
              "$ref": "#/components/schemas/Item"
            },
            "type": "array",
            "title": "Equipped"
          },
          "inventory": {
            "items": {
              "$ref": "#/components/schemas/Item"
            },
            "type": "array",
            "title": "Inventory"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "CharacterDetail"
      },
      "CharacterSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "realm": {
            "type": "string",
            "title": "Realm",
            "default": "pc"
          },
          "class": {
            "type": "string",
            "title": "Class"
          },
          "level": {
            "type": "integer",
            "title": "Level",
            "default": 0
          },
          "league": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "League"
          },
          "experience": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "class"
        ],
        "title": "CharacterSummary"
      },
      "CharactersResponse": {
        "properties": {
          "league": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "League"
          },
          "characters": {
            "items": {
              "$ref": "#/components/schemas/CharacterSummary"
            },
            "type": "array",
            "title": "Characters"
          }
        },
        "type": "object",
        "required": [
          "league",
          "characters"
        ],
        "title": "CharactersResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Item": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "inventory_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inventory Id"
          },
          "w": {
            "type": "integer",
            "title": "W",
            "default": 1
          },
          "h": {
            "type": "integer",
            "title": "H",
            "default": 1
          },
          "x": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "X"
          },
          "y": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "type_line": {
            "type": "string",
            "title": "Type Line",
            "default": ""
          },
          "base_type": {
            "type": "string",
            "title": "Base Type",
            "default": ""
          },
          "rarity": {
            "type": "string",
            "title": "Rarity",
            "default": "Normal"
          },
          "ilvl": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ilvl"
          },
          "identified": {
            "type": "boolean",
            "title": "Identified",
            "default": true
          },
          "corrupted": {
            "type": "boolean",
            "title": "Corrupted",
            "default": false
          },
          "properties": {
            "items": {
              "$ref": "#/components/schemas/ItemProperty"
            },
            "type": "array",
            "title": "Properties"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/ItemProperty"
            },
            "type": "array",
            "title": "Requirements"
          },
          "implicit_mods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Implicit Mods"
          },
          "explicit_mods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Explicit Mods"
          },
          "rune_mods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rune Mods"
          },
          "enchant_mods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Enchant Mods"
          },
          "crafted_mods": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Crafted Mods"
          },
          "sockets": {
            "items": {
              "$ref": "#/components/schemas/Socket"
            },
            "type": "array",
            "title": "Sockets"
          },
          "stack_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stack Size"
          },
          "max_stack_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Stack Size"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "raw": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "Item"
      },
      "ItemProperty": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ItemProperty"
      },
...
      "PrefsPatch": {
        "properties": {
          "trade_tolerance_pct": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 500.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Tolerance Pct"
          },
          "preferred_league": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred League"
          },
          "valuable_threshold_chaos": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Valuable Threshold Chaos"
          }
        },
        "type": "object",
        "title": "PrefsPatch"
      },
      "PrefsResponse": {
        "properties": {
          "trade_tolerance_pct": {
            "type": "integer",
            "title": "Trade Tolerance Pct"
          },
          "preferred_league": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred League"
          },
          "valuable_threshold_chaos": {
            "type": "integer",
            "title": "Valuable Threshold Chaos"
          }
        },
        "type": "object",
        "required": [
          "trade_tolerance_pct",
          "preferred_league",
          "valuable_threshold_chaos"
        ],
        "title": "PrefsResponse"
      },
...
      "Socket": {
        "properties": {
          "group": {
            "type": "integer",
            "title": "Group",
            "default": 0
          },
          "type": {
            "type": "string",
            "title": "Type",
            "default": ""
          }
        },
        "type": "object",
        "title": "Socket"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}